+++ src/Products/PloneMeeting/tests/testViews.py
@@ -764,6 +764,59 @@ class testViews(PloneMeetingTestCase):
             helper.printXhtml(item, [motivation, '<br>'], use_safe_html=False),
             motivation + '<br>')
 
+    def test_pm_PrintAdvicesInfos(self):
+        """Test the printAdvicesInfos method."""
+        cfg = self.meetingConfig
+        cfg.setSelectableAdvisers((self.developers_uid, self.vendors_uid))
+        cfg.setItemAdviceStates((self._stateMappingFor('itemcreated'),))
+        cfg.setItemAdviceEditStates((self._stateMappingFor('itemcreated'),))
+        cfg.setItemAdviceViewStates((self._stateMappingFor('itemcreated'),))
+        cfg.at_post_edit_script()
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem')
+        item.setOptionalAdvisers((self.developers_uid, self.vendors_uid), )
+        item._update_after_edit()
+        view = item.restrictedTraverse('document-generation')
+        helper = view.get_generation_context_helper()
+
+        # advices not given
+        self.assertEqual(
+            helper.printAdvicesInfos(item),
+            "<p class='pmAdvices'><u><b>Advices :</b></u></p>"
+            "<p class='pmAdvices'><u>Developers:</u><br /><u>Advice type :</u> "
+            "<i>Not given yet</i></p><p class='pmAdvices'><u>Vendors:</u><br />"
+            "<u>Advice type :</u> <i>Not given yet</i></p>")
+        self.changeUser('pmAdviser1')
+        createContentInContainer(item,
+                                 'meetingadvice',
+                                 **{'advice_group': self.developers_uid,
+                                    'advice_type': u'positive',
+                                    'advice_comment': RichTextValue(u'My comment')})
+        # mixes advice given and not given
+        self.assertEqual(
+            helper.printAdvicesInfos(item),
+            "<p class='pmAdvices'><u><b>Advices :</b></u></p>"
+            "<p class='pmAdvices'><u>Vendors:</u><br /><u>Advice type :</u> "
+            "<i>Not given yet</i></p><p class='pmAdvices'><u>Developers:</u><br />"
+            "<u>Advice type :</u> <i>Positive</i><br /><u>Advice given by :</u> "
+            "<i>M. PMAdviser One</i><br /><u>Advice comment :</u> My comment<p></p></p>")
+
+        # every advices given
+        self.changeUser('pmReviewer2')
+        createContentInContainer(item,
+                                 'meetingadvice',
+                                 **{'advice_group': self.vendors_uid,
+                                    'advice_type': u'negative'})
+        self.changeUser('pmCreator1')
+        self.assertEqual(
+            helper.printAdvicesInfos(item),
+            "<p class='pmAdvices'><u><b>Advices :</b></u></p><p class='pmAdvices'>"
+            "<u>Vendors:</u><br /><u>Advice type :</u> <i>Negative</i><br />"
+            "<u>Advice given by :</u> <i>M. PMReviewer Two</i><br />"
+            "<u>Advice comment :</u> -<p></p></p><p class='pmAdvices'><u>Developers:</u><br />"
+            "<u>Advice type :</u> <i>Positive</i><br /><u>Advice given by :</u> "
+            "<i>M. PMAdviser One</i><br /><u>Advice comment :</u> My comment<p></p></p>")
+
     def test_pm_MeetingUpdateItemReferences(self):
         """Test call to @@update-item-references from the meeting that will update
            every references of items of a meeting."""